+++ semcore/wizard/__tests__/wizard.browser-test.tsx
@@ -2,7 +2,7 @@ import { expect, test } from '@semcore/testing-utils/playwright';
 import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
 
 const locators = {
-  trigger: (page: any, name = 'Open modal') => page.getByRole('button', { name }),
+  trigger: (page: any, name = 'Open wizard') => page.getByRole('button', { name }),
   modal: (page: any) => page.locator("div[data-ui-name='Wizard']"),
   sidebar: (page: any) => page.locator("[data-ui-name='Wizard.Sidebar']"),
   stepperTabs: (page: any) => page.locator("[data-ui-name='Wizard.Stepper']"),
@@ -408,11 +408,11 @@ test.describe('Custom stepper example', () => {
     });
 
     await test.step('Click inside text fields and check focus', async () => {
-      await input(page, 'Your name').click();
-      await expect(input(page, 'Your name')).toBeFocused();
-      await input(page, 'Your name').fill('Test');
-      await input(page, 'Your email').click();
-      await expect(input(page, 'Your email')).toBeFocused();
+      await input(page, 'Name').click();
+      await expect(input(page, 'Name')).toBeFocused();
+      await input(page, 'Name').fill('Test');
+      await input(page, 'Email').click();
+      await expect(input(page, 'Email')).toBeFocused();
     });
 
     await test.step('Verify navigation between pages by pressing Steps in sidebar ', async () => {