+++ stories/components/data-table/docs/examples/column-expand.tsx
@@ -3,7 +3,7 @@ import DataTable from '@semcore/data-table';
 
 const Demo = () => {
   return (
-    <DataTable data={data} aria-label={'Table title. Column expanded'}>
+    <DataTable data={data} aria-label={'Column expanded'}>
       <DataTable.Head>
         <DataTable.Column name='keyword' children='Keyword' wMax={'300px'} />
         <DataTable.Column name='kd' children='Difficulty Difficulty' wMax={'85px'} />