+++ stories/components/inline-edit/docs/examples/simple_use.tsx
@@ -11,7 +11,9 @@ const Example = () => {
 
   return (
     <div>
-      <Text mr={2}>Author:</Text>
+      <Text mr={2} id='author-label'>
+        Author:
+      </Text>
       <InlineEdit editable={editable} onEditableChange={setEditable}>
         <InlineEdit.View style={{ display: 'flex', gap: 8, alignItems: 'center' }} pr={2}>
           {text} <EditM color='icon-secondary-neutral' />
@@ -28,7 +30,12 @@ const Example = () => {
             }}
             onBlurBehavior={'confirm'}
           >
-            <InlineInput.Value autoFocus value={text} onChange={setText} aria-label='Author name' />
+            <InlineInput.Value
+              autoFocus
+              value={text}
+              onChange={setText}
+              aria-labelledby='author-label'
+            />
             <InlineInput.ConfirmControl />
             <InlineInput.CancelControl />
           </InlineInput>