+++ stories/components/modal/advanced/modal.stories.tsx
@@ -1,9 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import ModalWithDatePickerExample from './examples/modal_open_dynamically';
+import ModalWithSelectExample from './examples/modal_with_select';
+import OpenModalDynamicallyExample from './examples/modal_open_dynamically';
 import ModalPreventFocusExample from './examples/modal_prevent_focus';
 import ModaliFrameExample from './examples/modal_iframe';
-
+import ModalIconExample from './examples/modal_with_int_icon';
 
 const meta: Meta = {
   title: 'Components/Modal/Advanced',
@@ -11,14 +12,22 @@ const meta: Meta = {
 
 export default meta;
 
-export const ModalWithDatePicker: StoryObj = {
-  render: ModalWithDatePickerExample,
+export const ModalWithSelect: StoryObj = {
+  render: ModalWithSelectExample,
+};
+
+export const OpenModalDynamically: StoryObj = {
+  render: OpenModalDynamicallyExample,
 };
 
 export const ModalPreventFocus: StoryObj = {
-    render: ModalPreventFocusExample,
-  };
-  
-  export const ModaliFrame: StoryObj = {
-    render: ModaliFrameExample,
-  };
+  render: ModalPreventFocusExample,
+};
+
+export const ModaliFrame: StoryObj = {
+  render: ModaliFrameExample,
+};
+
+export const ModalIcon: StoryObj = {
+  render: ModalIconExample,
+};