+++ stories/components/select/docs/examples/dropdownmenu_customization.tsx
@@ -17,11 +17,11 @@ const noticeStyle = {
 const Demo = () => (
   <Flex direction='column'>
     <Text tag='label' size={200} htmlFor='customized-dropdown-select'>
-      Customized dropdown
+      Select with notice
     </Text>
-    <Select placeholder={'Select something'}>
+    <Select>
       <Select.Trigger mt={2} mr='auto' id='customized-dropdown-select' />
-      <Select.Popper aria-label={'Options and notice'}>
+      <Select.Popper aria-label={'Select with notice'}>
         <Select.List hMax='240px'>
           {options.map((option, index) => (
             <Select.Option value={option} key={index}>