+++ stories/components/wizard/docs/examples/custom_step.tsx
@@ -12,10 +12,10 @@ const Step1 = React.forwardRef(function (_props, ref: React.Ref<HTMLDivElement>)
         Keywords
       </Text>
       <Input>
-        <Input.Value placeholder='Keyword 1' />
+        <Input.Value placeholder='Keyword 1' aria-label='Keyword 1' />
       </Input>
       <Input>
-        <Input.Value placeholder='Keyword 2' />
+        <Input.Value placeholder='Keyword 2' aria-label='Keyword 2' />
       </Input>
     </Flex>
   );
@@ -36,7 +36,7 @@ const Demo = () => {
 
   return (
     <>
-      <Button onClick={handleOpen}>Open modal</Button>
+      <Button onClick={handleOpen}>Open wizard</Button>
       <Wizard visible={visible} step={step} w={600} h={400} onClose={handleClose}>
         <Wizard.Sidebar title='Site Audit Settings'>
           <Wizard.Stepper step={1} onActive={handleStepChange(1)}>