+++ website/docs/utils/popper/examples/click-outside.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import Popper from 'intergalactic/popper';
+import Popper from '@semcore/popper';
+import Button from '@semcore/button';
 
 const style = { background: '#FFF', color: '#000', border: '1px solid #000', padding: '10px' };
 
@@ -10,7 +11,7 @@ const Demo = () => (
       return false;
     }}
   >
-    <Popper.Trigger style={style}>Click me pls</Popper.Trigger>
+    <Popper.Trigger tag={Button}>Open popper</Popper.Trigger>
     <Popper.Popper style={style}>Attached content</Popper.Popper>
   </Popper>
 );