+++ semcore/base-trigger/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.51.1] - 2025-02-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/flex-box` [5.40.0 ~> 5.40.1], `@semcore/utils` [4.45.0 ~> 4.45.1]).
+
+## [4.51.0] - 2025-02-05
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/icon` [4.56.0 ~> 4.57.0]).
+
+## [4.50.0] - 2025-02-03
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/utils` [4.44.1 ~> 4.45.0]).
+
 ## [4.49.4] - 2025-01-21
 
 ### Changed