+++ semcore/color-picker/CHANGELOG.md
@@ -8,6 +8,24 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 * `Item` element for `ColorPicker.Colors` and `PaletteManager.Colors`.
 
+## [2.42.2] - 2025-02-05
+
+### Changed
+
+* Version patch update due to children dependencies update (`@semcore/button` [5.41.1 ~> 5.41.2], `@semcore/flex-box` [5.40.0 ~> 5.40.1], `@semcore/utils` [4.45.0 ~> 4.45.1]).
+
+## [2.42.1] - 2025-02-05
+
+### Changed
+
+* Version patch update due to children dependencies update (`@semcore/button` [5.41.0 ~> 5.41.1], `@semcore/icon` [4.56.0 ~> 4.57.0]).
+
+## [2.42.0] - 2025-02-03
+
+### Changed
+
+* Version minor update due to children dependencies update (`@semcore/button` [5.40.3 ~> 5.41.0], `@semcore/utils` [4.44.1 ~> 4.45.0]).
+
 ## [2.41.8] - 2025-01-21
 
 ### Changed