+++ semcore/d3-chart/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.58.7] - 2025-02-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/animation` [2.40.0 ~> 2.40.1], `@semcore/flex-box` [5.40.0 ~> 5.40.1], `@semcore/utils` [4.45.0 ~> 4.45.1], `@semcore/core` [2.38.0 ~> 2.38.1]).
+
+## [3.58.6] - 2025-02-03
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/animation` [2.39.1 ~> 2.40.0], `@semcore/popper` [5.45.1 ~> 5.45.2], `@semcore/utils` [4.44.1 ~> 4.45.0], `@semcore/core` [2.37.1 ~> 2.38.0]).
+
 ## [3.58.5] - 2024-12-20
 
 ### Added