+++ semcore/i18n-unplugin/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [0.18.17] - 2025-02-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/core` [2.38.0 ~> 2.38.1]).
+
+## [0.18.16] - 2025-02-03
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/core` [2.37.1 ~> 2.38.0]).
+
 ## [0.18.15] - 2024-12-30
 
 ### Changed