+++ semcore/tooltip/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [6.47.3] - 2025-02-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.45.0 ~> 4.45.1], `@semcore/flex-box` [5.40.0 ~> 5.40.1], `@semcore/popper` [5.45.2 ~> 5.45.3], `@semcore/portal` [3.40.0 ~> 3.40.1], `@semcore/core` [2.38.0 ~> 2.38.1]).
+
+## [6.47.2] - 2025-02-03
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.44.1 ~> 4.45.0], `@semcore/flex-box` [5.39.1 ~> 5.40.0], `@semcore/popper` [5.45.1 ~> 5.45.2], `@semcore/portal` [3.39.1 ~> 3.40.0], `@semcore/core` [2.37.1 ~> 2.38.0]).
+
 ## [6.47.1] - 2024-12-16
 
 ### Changed